# cloud/openapi/shakecloud.yaml
openapi: 3.1.0
...
info:
  title: shake-cloud API
  version: 0.1.0
  description: |
    Login, access keys, the audit log and compute. Instances can be created
    with free-form cpu, memory and disk; named instance types are only a
    shorthand starting point, not a fixed catalogue of sizes. Quotas and
    capacity limits are administrator-settable at runtime through
    /v1/limits, not baked into the deployment.

    Browsers log in through Authentik and hold a session cookie. Terraform
    and the CLI send `Authorization: Bearer sca_<access_key_id>.<secret>`.
    Keys work while Authentik is down.
servers:
  - url: http://192.168.10.205:8080
security:
  - accessKey: []
  - session: []
...
paths:
...
  /v1/instance-types:
    get:
      operationId: DescribeInstanceTypes
      tags: [compute]
      summary: List the instance sizes, smallest first.
      description: The same names Terraform uses in platform/terraform/flavors.yaml.
      responses:
        '200':
          description: Instance types.
          content:
            application/json:
              schema:
                type: object
                required: [instance_types]
                properties:
                  instance_types:
                    type: array
                    items:
                      type: object
                      required: [instance_type, vcpus, memory_mib, memory_min_mib]
                      properties:
                        instance_type: {type: string}
                        vcpus: {type: integer}
                        memory_mib:
                          type: integer
                          description: The ceiling, and what quota and capacity count.
                        memory_min_mib:
                          type: integer
                          description: The floor the host may balloon down to.
        '401': {$ref: '#/components/responses/Error'}
...
  /v1/instances:
    post:
      operationId: RunInstances
      tags: [compute]
      summary: Launch an instance.
      description: |
        Returns immediately with the instance in `pending`; a worker then
        allocates the address, builds the seed ISO and creates the VM. Poll
        DescribeInstance until it is `running`, or read state_reason to see why
        it ended up `terminated`.

        Repeating a request with the same client_token returns the original
        instance (200) instead of launching another.

        **Sizing is free-form; instance_type is only a starting point.**
        Either `instance_type`, or both `vcpus` and `memory_mib`, must be
        given. Explicit `vcpus`, `memory_mib`, `memory_min_mib` or
        `root_disk_gib` override the preset's values.

        `ballooning` defaults to true. When true and no `memory_min_mib` is
        given, the floor defaults to a quarter of `memory_mib` (at least 512
        MiB). When false, the VM gets fixed memory and the balloon driver is
        disabled.

        vCPUs may not exceed the node's thread count; memory and disk are
        bounded by the account quota and the cloud's capacity checks, not by
        a fixed catalogue of sizes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                image_id:
                  type: string
                  pattern: '^img-[a-z0-9-]+$'
                  description: 'Launch by copying this image to the root disk. Give either image_id or install_iso_id, not both.'
                install_iso_id:
                  type: string
                  pattern: '^iso-[a-z0-9-]+$'
                  description: 'Install instead of launch: the VM boots this ISO with an empty root disk. Give either image_id or install_iso_id, not both.'
                driver_iso_id:
                  type: string
                  pattern: '^iso-[a-z0-9-]+$'
                  description: 'Optional second CD-ROM, e.g. the virtio-win driver disc, attached when install_iso_id is used.'
                guest_os:
                  type: string
                  enum: ['', linux, windows]
                  description: 'Defaults from the image or ISO. windows selects UEFI, a TPM 2.0 and the q35 chipset.'
                instance_type:
                  type: string
                  description: A named preset from GET /v1/instance-types, used as a starting point.
                vcpus: {type: integer, minimum: 1}
                memory_mib: {type: integer, minimum: 512}
                memory_min_mib:
                  type: integer
                  minimum: 512
                  description: The balloon floor. Ignored when ballooning is false.
                ballooning: {type: boolean, default: true}
                root_disk_gib:
                  type: integer
                  description: Omit for the deployment's default. Never smaller than the image.
                user_data:
                  type: string
                  maxLength: 16384
                  description: |
                    cloud-init user data, passed verbatim on a CIDATA CD-ROM.
                    It is readable from inside the instance, so it is not a
                    place for secrets.
                client_token: {type: string, maxLength: 64}
                key_name:
                  type: string
                  maxLength: 64
                  description: 'Must be one of the caller''s own key pairs. Its public key is added to the instance''s ssh_authorized_keys through cloud-init.'
                tags:
                  type: object
                  additionalProperties: {type: string}
                  description: 'Name becomes the guest hostname. Keys starting with shakecloud: are reserved.'
                security_group_ids:
                  type: array
                  minItems: 1
                  maxItems: 5
                  items: {type: string, pattern: '^sg-[0-9a-f]{17}$'}
                  description: 'Must belong to the caller''s own account. Omitted means the account''s default group.'
      responses:
        '202':
          description: Accepted; the instance is pending.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstanceEnvelope'}
        '200':
          description: The client_token had already been used for this exact request.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstanceEnvelope'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '409':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
    get:
      operationId: DescribeInstances
      tags: [compute]
      summary: List instances, newest first.
      description: |
        Every authenticated caller sees every instance in the cloud, not only
        their own, so that users can see what the cloud is running and who
        owns it. client_token is returned only for instances the caller owns.
        Terminated instances stay listed for an hour.

        Listing does not grant control: acting on an instance (power,
        terminate) still requires owning it or being a cloud-admin.
      parameters:
        - name: account_id
          in: query
          description: Filter by account. Available to everyone, not only cloud-admins.
          schema: {type: string, pattern: '^[0-9]{12}$'}
      responses:
        '200':
          description: Instances.
          content:
            application/json:
              schema:
                type: object
                required: [instances]
                properties:
                  instances:
                    type: array
                    items: {$ref: '#/components/schemas/Instance'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
  /v1/instances/adopt:
    post:
      operationId: AdoptInstance
      tags: [compute]
      summary: Register a VM that already exists into the cloud. cloud-admins only.
      description: |
        cloud-admins only.

        Takes over a VM the administrator has already moved into the cloud
        pool. It is registered under `account_id` and from then on behaves like
        a launched instance: power, console, tags and security groups all work.
        It cannot be launched or re-created, so it keeps no source image or
        seed ISO.

        The VM must be running or stopped, have a net0 MAC address, and its
        root disk must be measurable; if storage does not report the size, pass
        `root_disk_gib`. An address is optional, but security groups filter by
        address, so an adopted VM with none cannot be filtered.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdoptInstanceRequest'}
      responses:
        '201':
          description: The adopted instance.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstanceEnvelope'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /v1/instances/{instance_id}:
    get:
      operationId: DescribeInstance
      tags: [compute]
      summary: Read one instance.
      parameters:
        - $ref: '#/components/parameters/InstanceID'
      responses:
        '200':
          description: The instance.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstanceEnvelope'}
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
    delete:
      operationId: TerminateInstance
      tags: [compute]
      summary: Terminate an instance.
      description: |
        Returns immediately with the instance `shutting-down`. The VM, its disk,
        the seed ISO and the address are all released before it reaches
        `terminated`. Terminating a terminated instance is not an error.
      parameters:
        - $ref: '#/components/parameters/InstanceID'
      responses:
        '202':
          description: Accepted.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstanceEnvelope'}
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
    patch:
      operationId: ModifyInstance
      tags: [compute]
      summary: Change an instance's cpu, memory, ballooning or disk size. cloud-admins only.
      description: |
        cloud-admins only.

        `vcpus`, `memory_mib` and `ballooning` can only be changed while the
        instance is **stopped**; Proxmox applies them at the next start.
        `root_disk_gib` may be grown while the instance runs. A disk can only
        grow, never shrink.

        The new size is re-checked against the account quota and the host's
        capacity, so a change can be refused for the same reasons a launch
        can.
      parameters:
        - $ref: '#/components/parameters/InstanceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              description: At least one property is required.
              properties:
                vcpus: {type: integer, minimum: 1}
                memory_mib: {type: integer, minimum: 512}
                memory_min_mib: {type: integer, minimum: 512}
                ballooning: {type: boolean}
                root_disk_gib: {type: integer}
      responses:
        '200':
          description: The modified instance.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstanceEnvelope'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
...
components:
  parameters:
    InstanceID:
      name: instance_id
      in: path
      required: true
      schema: {type: string, pattern: '^i-[0-9a-f]{17}$'}
...
  securitySchemes:
    accessKey:
      type: http
      scheme: bearer
      bearerFormat: sca_<access_key_id>.<secret>
    session:
      type: apiKey
      in: cookie
      name: shakecloud_session
  responses:
    Error:
      description: |
        Codes: AuthFailure (401), UnauthorizedOperation (403),
        CrossOriginRequestBlocked (403), NoSuchEntity (404), ValidationError,
        InvalidParameterValue and MalformedRequest (400), IncorrectInstanceState
        and LimitExceeded (409), InternalError (500), ServiceUnavailable (503).
        Also InvalidImageID.NotFound (404), ImageLimitExceeded (409),
        RequestEntityTooLarge (413), InvalidKeyPair.Duplicate (409) and
        InvalidKeyPair.NotFound (404).
        Volumes and security groups add InvalidVolume.NotFound (404),
        IncorrectState (409), VolumeInUse (409), VolumeLimitExceeded (409),
        AttachmentLimitExceeded (409), InvalidInstanceID.NotFound (404),
        IncorrectInstanceState (409), IdempotentParameterMismatch (409),
        InvalidGroup.NotFound (404), InvalidGroup.Duplicate (409),
        InvalidPermission.Duplicate (409), InvalidPermission.NotFound (404),
        DependencyViolation (409), CannotDelete (409),
        InsufficientVolumeCapacity (503),
        RulesPerSecurityGroupLimitExceeded (409),
        SecurityGroupsPerInstanceLimitExceeded (409) and
        SecurityGroupLimitExceeded (409).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
        request_id: {type: string}
...
    InstanceEnvelope:
      type: object
      required: [instance]
      properties:
        instance: {$ref: '#/components/schemas/Instance'}
    Instance:
      type: object
      required: [instance_id, account_id, image_id, state, mac_address, vcpus, memory_mib, root_disk_gib, ballooning, launch_time, security_groups, firewall_state]
      properties:
        instance_id: {type: string, pattern: '^i-[0-9a-f]{17}$'}
        account_id: {type: string}
        owner_username:
          type: string
          description: The account's username, so a viewer can tell whose instance this is.
        image_id:
          type: string
          description: Empty for an instance that installs itself from ISO media instead of launching from an image.
        image_name: {type: string, description: The human name of the image.}
        guest_os: {type: string, enum: ['', linux, windows], description: 'windows for a Windows guest. Empty for a Linux guest.'}
        install_iso_id: {type: string, description: 'Set when the instance was created from install media.'}
        driver_iso_id: {type: string, description: 'The optional second CD-ROM, e.g. the virtio-win disc.'}
        instance_type:
          type: string
          description: Absent when the instance was created with explicit cpu/memory rather than a named preset.
        state:
          type: string
          enum: [pending, running, stopping, stopped, shutting-down, terminated]
        pending_action:
          type: string
          enum: [launch, start, stop, reboot, terminate]
          description: The lifecycle action a worker is still carrying out; absent when the instance is settled.
        state_reason:
          type: string
          description: Why the instance is in this state, when it was not asked for.
        private_ip_address:
          type: string
          description: Set once the address has been allocated.
        mac_address: {type: string}
        vcpus: {type: integer}
        memory_mib:
          type: integer
          description: The ceiling. What quota and the memory budget count.
        memory_min_mib:
          type: integer
          description: The floor the host may reclaim this instance's memory down to. 0 when ballooning is off.
        ballooning:
          type: boolean
          description: Whether the memory balloon driver is enabled. When false the instance has fixed memory and memory_min_mib is 0.
        root_disk_gib: {type: integer}
        tags:
          type: object
          additionalProperties: {type: string}
        client_token: {type: string}
        key_name: {type: string, description: 'Absent when no key was used.'}
        launch_time: {type: string, format: date-time}
        terminated_at: {type: string, format: date-time}
        security_groups:
          type: array
          items: {$ref: '#/components/schemas/InstanceSecurityGroup'}
          description: 'Empty for instances launched before security groups existed; such instances are unfiltered.'
        firewall_state:
          type: string
          enum: [in-sync, applying]
          description: Whether the last change to security_groups has finished being pushed to the guest's firewall rules.
        adopted:
          type: boolean
          description: 'True for a VM that already existed and was registered into the cloud, rather than created by a launch. It has no source image.'
    AdoptInstanceRequest:
      type: object
      additionalProperties: false
      required: [vmid, account_id]
      properties:
        vmid:
          type: integer
          minimum: 1
          description: An existing VM in the cloud pool on this node.
        account_id:
          type: string
          pattern: '^[0-9]{12}$'
          description: The owner the administrator chooses. The VM's quota is checked against this account.
        name:
          type: string
          maxLength: 255
          description: Omit to keep the VM's own name.
        private_ip_address:
          type: string
          description: Optional. Security groups filter by address, so an instance with none cannot be filtered.
        root_disk_gib:
          type: integer
          minimum: 1
          description: Omit to measure the disk from storage; pass it when storage does not report the size.
        tags:
          type: object
          additionalProperties: {type: string}
        security_group_ids:
          type: array
          minItems: 1
          maxItems: 5
          items: {type: string, pattern: '^sg-[0-9a-f]{17}$'}
          description: 'Must belong to the chosen account. Omitted means the account''s default group.'
...
    InstanceSecurityGroup:
      type: object
      required: [group_id, group_name]
      properties:
        group_id: {type: string, pattern: '^sg-[0-9a-f]{17}$'}
        group_name: {type: string}
...